% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-MulNIRAdata.R
\docType{data}
\name{MulNIRAdata}
\alias{MulNIRAdata}
\title{Multilevel NIRA Example Data}
\format{
A data frame with 1000 rows and 8 columns, where:
\describe{
\item{NA,UW,WTM, …, ASH}{Symptoms in GAD-7,Binary variables (0/1) serving as network nodes.}
\item{city}{Factor indicating the grouping variable (city).}
}
}
\source{
Preprocessed from original PBICR-2023 Anxiety data.
}
\usage{
data(MulNIRAdata)
}
\description{
A subset of the \verb{PBICR-2023 Anxiety data} dataset (first 1000 rows) used for demonstrating
the multilevel Ising-network fitting functions in this package.
}
\keyword{datasets}
