% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholds_select.R
\name{Thresholds_select}
\alias{Thresholds_select}
\title{Extract Thresholds from a Multilevel Ising Fit}
\usage{
Thresholds_select(fit, type = c("random", "fixed"), group = NULL, level = NULL)
}
\arguments{
\item{fit}{The object returned by \code{Multilevel_Isingfit()}, containing the fitted multilevel Ising model results.}

\item{type}{A string specifying the intercept type to extract: \code{"random"} for random intercepts or \code{"fixed"} for fixed intercepts.}

\item{group}{A character string naming the grouping variable; required when \code{type = "random"}. This should match the \code{group.col} argument used in \code{Multilevel_Isingfit()}.}

\item{level}{An integer indicating which level’s random intercept to extract; required when \code{type = "random"}. For example, \code{level = 5} refers to the fifth hierarchical level.}
}
\description{
Retrieve fixed or random‑intercept thresholds from an object of class \code{IsingFit_ml}.
}
\examples{
Thresholds_select(Mul_Isingmodel, type = "fixed")
Thresholds_select(Mul_Isingmodel,type  = "random",group = "city",level = 5)
}
\author{
Written by Wu Yiming 2025/07/29
MultilevelNIRA
School of Psychology, South China Normal University
}
